server/requirements.txt
Django==5.0.4
psycopg2-binary==2.9.6
gunicorn==21.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0

django-colorfield==0.12.0
django-durationwidget==1.0.5

debugpy==1.8.12